Vertical color legend (#347)
