+++ .changeset/dasd_ddasf_asdasd.md
@@ -1,0 +1,5 @@
+---
+'@ldn-viz/ui': major
+---
+
+CHANGED: rename the `ColorLegendOrdinalHorizontal` component to `ColorLegendOrdinalChips`, and allos `orientation` to be set to `vertical`
